Fix makefiles, all.sh scripts, templates, etc
It's large but I don't feel like explaining today :D

+++ problems/leet/388-longest-absolute-file-path/code.cpp
@@ -21,7 +21,7 @@ public:
           dot = true;
         ++namelen, ++i;
       }
-      if (lengths.size() < tabs + 1)
+      if (lengths.size() < size_t(tabs + 1))
         lengths.resize(tabs + 1, 0);
       lengths[tabs] = namelen;
       if (tabs > 0)
